# openapi-schema.yml
openapi: 3.0.3
info:
  title: SyncScope Alerts Service API
  version: 1.0.0
  description: Real-time alerting and notification service for SyncScope platform
paths:
...
  /alerts/:
    get:
      operationId: alerts_list
      description: ViewSet for managing alerts
      parameters:
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      tags:
      - alerts
      security:
      - Bearer: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedAlertList'
          description: ''
    post:
      operationId: alerts_create
      description: ViewSet for managing alerts
      tags:
      - alerts
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlertRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/AlertRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/AlertRequest'
        required: true
      security:
      - Bearer: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Alert'
          description: ''
  /alerts/{id}/:
    get:
      operationId: alerts_retrieve
      description: ViewSet for managing alerts
      parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
      tags:
      - alerts
      security:
      - Bearer: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertDetail'
          description: ''
    put:
      operationId: alerts_update
      description: ViewSet for managing alerts
      parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
      tags:
      - alerts
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlertRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/AlertRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/AlertRequest'
        required: true
      security:
      - Bearer: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Alert'
          description: ''
    patch:
      operationId: alerts_partial_update
      description: ViewSet for managing alerts
      parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
      tags:
      - alerts
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedAlertRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedAlertRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedAlertRequest'
      security:
      - Bearer: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Alert'
          description: ''
    delete:
      operationId: alerts_destroy
      description: ViewSet for managing alerts
      parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
      tags:
      - alerts
      security:
      - Bearer: []
      responses:
        '204':
          description: No response body
...
components:
  schemas:
    Alert:
      type: object
      description: Serializer for Alert model
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        rule:
          type: string
          format: uuid
        rule_name:
          type: string
          readOnly: true
        triggered_for_user_id:
          type: string
          format: uuid
          nullable: true
        triggered_for_team_id:
          type: string
          format: uuid
          nullable: true
        severity:
          type: string
          maxLength: 20
        title:
          type: string
          maxLength: 255
        message:
          type: string
        is_read:
          type: boolean
        acknowledged_by:
          type: string
          format: uuid
          nullable: true
        acknowledged_at:
          type: string
          format: date-time
          nullable: true
        triggered_at:
          type: string
          format: date-time
          readOnly: true
        status:
          type: string
          maxLength: 20
        context_data: {}
      required:
      - id
      - message
      - rule
      - rule_name
      - severity
      - title
      - triggered_at
    AlertDetail:
      type: object
      description: Detailed serializer for Alert with full alert rule info
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        rule:
          allOf:
          - $ref: '#/components/schemas/AlertRule'
          readOnly: true
        rule_name:
          type: string
          readOnly: true
        triggered_for_user_id:
          type: string
          format: uuid
          nullable: true
        triggered_for_team_id:
          type: string
          format: uuid
          nullable: true
        severity:
          type: string
          maxLength: 20
        title:
          type: string
          maxLength: 255
        message:
          type: string
        is_read:
          type: boolean
        acknowledged_by:
          type: string
          format: uuid
          nullable: true
        acknowledged_at:
          type: string
          format: date-time
          nullable: true
        triggered_at:
          type: string
          format: date-time
          readOnly: true
        status:
          type: string
          maxLength: 20
        context_data: {}
      required:
      - id
      - message
      - rule
      - rule_name
      - severity
      - title
      - triggered_at
    AlertRequest:
      type: object
      description: Serializer for Alert model
      properties:
        rule:
          type: string
          format: uuid
        triggered_for_user_id:
          type: string
          format: uuid
          nullable: true
        triggered_for_team_id:
          type: string
          format: uuid
          nullable: true
        severity:
          type: string
          minLength: 1
          maxLength: 20
        title:
          type: string
          minLength: 1
          maxLength: 255
        message:
          type: string
          minLength: 1
        is_read:
          type: boolean
        acknowledged_by:
          type: string
          format: uuid
          nullable: true
        acknowledged_at:
          type: string
          format: date-time
          nullable: true
        status:
          type: string
          minLength: 1
          maxLength: 20
        context_data: {}
      required:
      - message
      - rule
      - severity
      - title
    AlertRule:
      type: object
      description: Serializer for AlertRule model
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        name:
          type: string
          maxLength: 255
        description:
          type: string
          nullable: true
        metric_type:
          type: string
          maxLength: 100
        condition:
          type: string
          maxLength: 50
        threshold_value:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
          nullable: true
        is_active:
          type: boolean
        check_interval_minutes:
          type: integer
        company_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - company_id
      - condition
      - created_at
      - id
      - metric_type
      - name
      - updated_at
...
    PaginatedAlertList:
      type: object
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/Alert'
...
    PatchedAlertRequest:
      type: object
      description: Serializer for Alert model
      properties:
        rule:
          type: string
          format: uuid
        triggered_for_user_id:
          type: string
          format: uuid
          nullable: true
        triggered_for_team_id:
          type: string
          format: uuid
          nullable: true
        severity:
          type: string
          minLength: 1
          maxLength: 20
        title:
          type: string
          minLength: 1
          maxLength: 255
        message:
          type: string
          minLength: 1
        is_read:
          type: boolean
        acknowledged_by:
          type: string
          format: uuid
          nullable: true
        acknowledged_at:
          type: string
          format: date-time
          nullable: true
        status:
          type: string
          minLength: 1
          maxLength: 20
        context_data: {}
...
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'Enter your Bearer token in the format: Bearer <token>'
servers:
- url: http://127.0.0.1:8003
  description: Local alerts service
- url: http://localhost:8003
  description: Local development server
- url: https://syncscope-alerts-service-dev.up.railway.app
  description: Development server
...